jama-to-ppt/requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
selenium>=4.15.0
webdriver-manager>=4.0.1

# PowerPoint Generation
python-pptx>=0.6.23

# Image Processing (optional)
Pillow>=10.1.0

# AI Features (optional)
anthropic>=0.18.0
